Change endianness of data on z80 stack. So LSW is on top of stack with MSW below it. This has advantage of making binary operations e.g. addition much more efficient as can add the LSW's together first before adding the MSW's.
Also have reversed the order of binary operator operands when loading on the stack again to make binary op code gen more efficient. So for v1 op v2 then v1 will be on top of the stack with v2 below it. This means the runtime arithmentic code can remove v1 in entirety e.g. into DEHL and then pop of v2 one word at a time using the words to implement the operation.

Change runtime routines to use pattern of popping return address into IY and then doing JP(IY) instead of RET at end of routine for better efficiency.

Fix bug in Snake when generating food - as was not hit testing for all of snake body but was only hit testing against head. Also init the xs/ys arrays to -1.