Refactor localloc to use a new evaluation stack node type of LocalHeapEntry.
This fixes bug with existing localloc that it will also allocate at least 4 bytes.

Also new code now adds cast from int32 to nativeint for the size to be allocated as we cannot allocate more than nativeint size of memory.

Note that it may be more efficient in some cases to allocate the memory as a temporary local but can revisit this later on.