Fix bug with BinaryOperators not having the correct exact size (#63)
Fix bug with Dup importing when value on top of evaluation stack is complex e.g. result of newarr then we can't simply duplicate this as would result in two newarr's. Change to use a spill to local and load local twice approach.
Add test to cover change to dup impoter.